+++ frontend/pages/demand-planning/index.tsx
@@ -5,9 +5,9 @@ import { Flex, HStack, Button, useDisclosure } from '@chakra-ui/react';
 
 import { apiClient } from '../../page-lib/demand-planning';
 
-import PlanList from '../../page-component/demand-planning/PlanList';
-import type { PlanData } from '../../page-component/demand-planning/Plan';
-import PlanForm from '../../page-component/demand-planning/PlanForm';
+import type { PlanData } from '../../page-component/demand-planning/Plan/Plan';
+import PlanList from '../../page-component/demand-planning/PlanList/PlanList';
+import PlanForm from '../../page-component/demand-planning/PlanForm/PlanForm';
 
 const DemandPlanning: NextPage = () => {
     const [plans, setPlans] = useState<PlanData[]>([]);
